Refactor notification_store canister structure and remove unused code
- Introduced a new `data_model` module and refactored the `CanisterData` structure for improved organization.
- Removed unused memory management code and notification handling functions to streamline the notification store canister.
- Updated imports in lifecycle files to reflect the new module structure.
